+++ Modules/luxon.js
@@ -1,4 +1,4 @@
-import { DateTime } from './luxonCript.js';
+import { DateTime } from './luxonFile.js';
 
 const dateTime = DateTime.now().toLocaleString(DateTime.DATETIME_MED);
