+++ Modules/selectors.js
@@ -1,0 +1,8 @@
+export const booklist = document.querySelector('#listOfBooks');
+export const bookTitle = document.querySelector('#bookTitle');
+export const bookAuthor = document.querySelector('#bookAuthor');
+export const addBook = document.querySelector('#addButton');
+export const navs = document.querySelectorAll('#navList > li');
+export const sections = document.querySelectorAll('section');
+export const addButton = document.querySelector('#addButton');
+export const date = document.querySelector('#date');
