+++ css/style.css
@@ -135,3 +135,10 @@ footer {
   flex-direction: column;
   gap: 15px;
 }
+.card-remove-button {
+  font-weight: bold;
+  margin: 7px;
+  margin-left: auto;
+  margin-right: 30px;
+  box-shadow: 4px 4px #000;
+}
